refactor: organise files
