+++ tailwind.config.ts
@@ -1,5 +1,5 @@
 import type { Config } from "tailwindcss";
-import { Themes } from "./theme";
+import { Themes } from "./misc/theme";
 
 const config: Config = {
   content: [
